refactor: improved readability and moved the code to cart-page-target
